refactor(plugin): package skeleton + util extraction + GE-tax dedup
Part of #729 (epic #728).

- Create com.flipsmart.util; move GpUtils, TimeUtils, ItemUtils, GeTax,
  BuyPriceLookup into it (pure, client-side helpers) and add the imports
  across all consumers.
- Centralize GE-tax math in GeTax: add price-only taxFor(int) and
  breakevenSellPrice(int), and route the six previously-duplicated tax
  sites (FocusedFlip, FlipAssistOverlay, FlipFinderPanel x3,
  GeOfferDescriptionFormatter) through it.
- Item-id-aware sites now correctly honor the GE tax-exempt item list;
  otherwise no behavior change.
