+++ src/main/java/com/flipsmart/AutoRecommendService.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.util.GpUtils;
 
 import com.flipsmart.FlipAssistOverlay.FlipAssistStep;
 import lombok.extern.slf4j.Slf4j;
