+++ src/main/java/com/flipsmart/GrandExchangeSlotOverlay.java
@@ -1,4 +1,8 @@
 package com.flipsmart;
+import com.flipsmart.util.BuyPriceLookup;
+import com.flipsmart.util.GeTax;
+import com.flipsmart.util.TimeUtils;
+import com.flipsmart.util.GpUtils;
 
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
