+++ src/main/java/com/flipsmart/GrandExchangeTracker.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.util.ItemUtils;
 
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.GrandExchangeOfferState;
