+++ src/main/java/com/flipsmart/ManualAdjustmentTracker.java
@@ -1,4 +1,5 @@
 package com.flipsmart;
+import com.flipsmart.util.GpUtils;
 
 import lombok.extern.slf4j.Slf4j;
